backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
aiomysql==0.2.0
alembic==1.13.1
python-dotenv==1.0.1
cryptography==42.0.5 
pandas==1.5.3
numpy==1.24.4